remove extraneous static asserts on transit functions; add asserts on illegal state
Remove static_assert on transit functions:

 - ActionFunction: remove check on void return type.
 - ConditionFunction: remove check on bool return type (implicit
   conversion!).

Note that the new static asserts are not strictly required, as the
assignment to current_state_ptr of different type fails anyway.
Gives comprehensive message on error.

Note: adding an abstraction for copy/paste implies:

    s/_state_instance<S>::value/state<S>()/

which is fine, but is left on purpose for simplicity / readability
reasons.