describe why we use no virtual-dtor
Building with `-Wnon-virtual-dtor` prints warnings for each
`tinyfsm::Fsm` class:

> warning: 'class Elevator' has virtual functions and accessible
> non-virtual destructor [-Wnon-virtual-dtor]

Adding a virtual destructor would waste resources (at least one vtable
entry). There is no reason to have a one, as the MooreMachine class
(and derived classes) is NOT intended to be deleted through a pointer
to it.

Building `examples/api/moore_machine.cpp` results in:

- without virtual destructor:

    # size -B -d moore_machine
       text    data     bss     dec     hex filename
       3277     768     584    4629    1215 moore_machine

- with virtual destructor in `MooreMachine`

    # size -B -d moore_machine
       text    data     bss     dec     hex filename
       3662     808     600    5070    13ce moore_machine

Ref: https://github.com/digint/tinyfsm/issues/21